Access the bytes of my media capture source
My laptop webcame returns 1843200 bytes with a stride of 2560. This represents a 720p frame, the resolution of my webcam. I need to figure out what format that represents, however, because the stride is certainly not RGB. Then I need to figure out how to write that into a WIC IBitmapSource.